Update README.md
This is a simple C++ Program that when executed will ask the user to input a filename. After the user enters a filename the program will then open the file and begin to iterate through each character and line of the file one letter at a time. The main function of this program is to count the number of words and how many times each word is repeated in the text file. It will then print a summary of all the words in the text file and how many times that word appeared. It will also print out at the bottom the total number of words and unique words in the file. The program has to account for spaces and special characters as well so it doesn't assume spaces, periods, or other grammatical notations are counted as words. This program uses the basics of C++ strings and basic fstream functions. While being a basic program, there is certain code that could be added to this program to make it more useful then it already is. There are multiple programs online that are similar to this one and just a few unique touches would put this one above them all.